+++ src/components/reducers/DatabaseDispatch.tsx
@@ -26,6 +26,7 @@ import { obtainUniqueOptionValues } from "helpers/SelectHelper";
 import { Literal } from "obsidian-dataview/lib/data-model/value";
 import { DateTime } from "luxon";
 import { RowSelectOption } from "cdm/ComponentsModel";
+import { viewport } from "@popperjs/core";
 
 export function databaseReducer(state: TableDataType, action: ActionType) {
   LOGGER.debug(`<=>databaseReducer action: ${action.type}`, action);
@@ -80,7 +81,8 @@ export function databaseReducer(state: TableDataType, action: ActionType) {
       VaultManagerDB.create_markdown_file(
         state.view.file.parent,
         action.filename,
-        rowRecord
+        rowRecord,
+        state.view.diskConfig.yaml.config
       );
       const metadata: Record<string, Literal> = {};
       metadata[MetadataColumns.CREATED] = DateTime.now();
@@ -104,6 +106,27 @@ export function databaseReducer(state: TableDataType, action: ActionType) {
       return update(state, {
         view: {
           rows: { $push: [row] },
+        },
+      });
+    /**
+     * Add new row into table
+     */
+    case ActionTypes.CHANGE_ROW_TEMPLATE:
+      state.view.diskConfig.updateConfig(
+        "current_row_template",
+        action.template
+      );
+      return update(state, {
+        view: {
+          diskConfig: {
+            yaml: {
+              config: {
+                current_row_template: {
+                  $set: action.template,
+                },
+              },
+            },
+          },
         },
       });
     /**